+++ src/main/java/team/po/feature/teamspace/domain/GithubInstallation.java
@@ -1,4 +1,4 @@
-package team.po.feature.projectgroup.domain;
+package team.po.feature.teamspace.domain;
 
 import java.time.Instant;
